backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
openai==0.27.8
anthropic==0.7.1
gunicorn==20.1.0
